refact : mock 데이터 mocks 디렉토리로 이동
